+++ src/index.ts
@@ -137,5 +137,5 @@ export interface HttpClient {
 
 export { makeHttpClient } from './core/http';
 
-export { LocalStorage, SessionStorage } from './core/storage';
 export { AsyncFileReader } from './core/file';
+export { LocalStorage, SessionStorage } from './core/storage';